Test Order addProduct() fail

+++ src/main/java/com/booleanuk/core/Coffee.java
@@ -2,7 +2,7 @@ package com.booleanuk.core;
 
 public class Coffee extends Product{
 
-    public Coffee(String sku, String variant, double price){
+    public Coffee(String sku, String variant, int price){
         super(sku, variant, price);
     }
 }
