+++ src/main/java/com/booleanuk/core/Filling.java
@@ -2,7 +2,7 @@ package com.booleanuk.core;
 
 public class Filling extends Product{
 
-    public Filling(String sku, String variant, double price){
+    public Filling(String sku, String variant, int price){
         super(sku, variant, price);
     }
 }
